@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Weekly::Application.config.secret_token = 'df1e15b6063f630b15976c51bfbe841af01e8b42de59ad32e75b88ebc84cd3de169a68beddfe64b04a94329fed42faa0a36b1496623cfff3cbe98d34578c3055'
+Weekly::Application.config.secret_token = 'af30b567672f8bd2e45c43070b68dce55e74e43afd8dabd68e6a4a08042dfc3bf18bdef86a30b861d2290bebb260cde2e80ab6e489e8d309444ab175447b563d'
