@@ -1,7 +1,8 @@
 Weekly::Application.routes.draw do
   resources :web_seeds
 
-  root to: "web_seeds#index" 
+  root to: "webSeeds#index"
+
 
   # The priority is based upon order of creation:
   # first created -> highest priority.
@@ -58,5 +59,5 @@ Weekly::Application.routes.draw do
 
   # This is a legacy wild controller route that's not recommended for RESTful applications.
   # Note: This route will make all actions in every controller accessible via GET requests.
-  # match ':controller(/:action(/:id(.:format)))'
+  # match ':controller(/:action(/:id))(.:format)'
 end
